[Decomp][Player] Cohesion split: extract attack-swing notifier senders to PlayerCombat.cpp
